// types/blue-server.ts
type middleware = (req: Request, res: Response, next: NextFunction) => any

/**
 * BlueServer types
 */
declare namespace BlueServer {
  /**
   * Blueserver fundamental configuration types
   */
  interface Config {
    /**
     * Port to listen on
     */
    port: number | string;
    /**
     * Host to listen on
     */
    host: string;
    /**
     * Cors options
     * 
     * See details: https://www.npmjs.com/package/cors
     */
    cors: CorsOptions;
    /**
     * API configurations
     * 
     */
    api: {
      base: string;
      /**
       * You can serve multiple static folders from the api.
       * 
       * See details: https://expressjs.com/en/4x/api.html#express.static
       */
      static: Static[];
      /**
        - Used by req.config
        - Generates the api with the given config
        - Can be extendable by middlewares
       */
      routes: {
        [name: string]: Route[]
      };
      /**
       * Definitions for the api routes
       * 
       */
      define: {
        /**
         * Order is important if there are foreign key references
         */
        models: Model.Tables;
        /**
          - Every fragment(column) is like a sub route used by the parent routes
          - Each parent column can be used for populating relational data or datasets from these sub routes
          - This parent column is usually an id or a virtual column
          - Example: the posts entity doesn't have column as `isLiked`, but you can populate it via `likes` table.
            A user token will be required for this, and likes table should have a relation with the post ids.
        */
        populate: Populate.Columns;
        utils?: object;
        /**
         * Middlewares for the api routes
         * 
         * See details: https://expressjs.com/en/guide/using-middleware.html
         */
        middlewares?: {
          [functionName: string]: middleware;
        }
      };
      plugins?: object;
    },
    /**
     * Knex database connection
     */
    database: any; // I will define this later.
    poweredBy?: string;
  }

  /**
   * Will transformed into the API routes
   */
  interface Route {
    /**
     * Route path
     * 
     * https://expressjs.com/en/guide/routing.html
     */
    path: string;
    /**
     * Route method
     * 
     */
    method: BlueServer.methods;
    /**
     * Controller option.
     * 
     * `upload` controller may have issues.
     */
    controller: BlueServer.controllers;
    /** 
     * Intercepts the request and carries response to the next segment.
     
     * You can define your functions inside of `api.define.middlewares` and call their name in this array.
     * 
     * Order is important.
     * 
     * You can also put your functions in this array instead of strings.
    */
    middlewares?: string | string[] | middleware | BlueServer.middlewares[];
    /**
     * Columns inherited from your model files.
     * 
     * This property represents selected columns for the response
     */
    columns?: string[];
    /**
     * It excludes specified columns from the columns array.
     */
    exclude?: string[];
    include?: string[];
    /**
     * Populates response data with the given columns.
     * 
     * Be sure that columns are defined.
     */
    populate?: string[];
    /**
     * Right now there is only one utility function.
     * 
     * Detect language, detects the language of the `content` column.
     * 
     * Column selection should be optional. Needs improvement.
     * */
    utils?: BlueServer.utils[];
    /**
     * Extends the query builder.
     * 
     * You can write your advanced sql queries and call via view.
     * 
     * If not defined, the controller name will be used.
     */
    view?: string;
  }
  /**
    - Used for creating a database
    - Used for api router and populate configurations
  */
  namespace Model {
    type types = "increments" | "integer" | "bigInteger" | "text" | "string" | "float" | "double" | "decimal" | "boolean" | "date" | "datetime" | "time" | "timestamp" | "timestamps" | "binary" | "enum" | "json" | "jsonb" | "uuid" | "geometry" | "geography" | "point";
    type constraints = "primary" | "nullable" | "notNullable" | "unique";
    interface Tables {
      [tableName: string]: {
        [columnName: string]: Properties
      }
    }
// ...
  }

  namespace Populate {
    interface Columns {
      [columnName: string]: Properties
    }
// ...
  }

  interface Static {
    /**
    * Virtual path of your static folder
    */
    base?: string;
    /**
     * Folder name, relative to root config file.
     * 
     * You can also define `fullpath` instead.
     */
    folder?: string;
    /**
     * Physical path of your static folder.
     */
    fullpath?: string;
    options: ServeStaticOptions;
  }
  type utils = "detect-language"
  type middlewares = "authorization" | "is-owner"
  type controllers = "find" | "findOne" | "count" | "delete" | "update" | "create" | "upload"
  type methods = "get" | "post" | "put" | "delete"
}